pyproject.toml
[project]
name = "codeshield-backend"
version = "0.1.0"
description = "CodeShield AI - AI-Powered Code Security Analysis Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.12",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.38",
    "uvicorn==0.34.0",
    "redis==5.0.0",
    "pydantic>=2.0.2,<3.0.0",
    "pydantic[email]>=2.0.2,<3.0.0",
    "google-generativeai==0.3.2",
    "httpx==0.28.1",
    "python-dateutil==2.8.2",
    "pytz==2023.3"
]


